requirements.txt
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
shap
streamlit
fastapi
uvicorn
google-generativeai
pytest
matplotlib
seaborn
python-multipart
Requests
python-dotenv
